# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Portal API",
        "description": "API for News Portal",
        "contact": {
            "name": "Your Name",
            "email": "your@email.com"
        },
        "license": {
            "name": "Your License",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/favorite/create": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add an article to the user's favorites list",
                "operationId": "b37e1220a837f2f716c2fe7fbfa5afe8",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "description": "The URL of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "The author of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "A brief description of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imageUrl",
                        "in": "query",
                        "description": "URL of the article's image",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article added to favorites successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Article added to favorites!"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, article already in favorites",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "This article is already in your favorites list!"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, user not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Authentication required"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Internal Server Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/favorite/list": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Get the list of user's favorite articles",
                "operationId": "e79eb2acb501979a41fa42b094069003",
                "responses": {
                    "200": {
                        "description": "List of user's favorite articles",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "favorites": [
                                        {
                                            "id": 1,
                                            "title": "Sample Article 1",
                                            "url": "http://sample-article-1.com",
                                            "author": "John Doe",
                                            "description": "Lorem ipsum dolor sit amet.",
                                            "imageUrl": "http://sample-article-1.com/image.jpg"
                                        },
                                        {
                                            "id": 2,
                                            "title": "Sample Article 2",
                                            "url": "http://sample-article-2.com",
                                            "author": "Jane Doe",
                                            "description": "Consectetur adipiscing elit.",
                                            "imageUrl": "http://sample-article-2.com/image.jpg"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, user not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Authentication required"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Internal Server Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/favorite/delete/{favorite}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Delete a favorite article",
                "operationId": "28aa003a86fef84eb3a0b77867ea8753",
                "parameters": [
                    {
                        "name": "favorite",
                        "in": "path",
                        "description": "ID of the favorite article to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article removed from favorites successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "Article removed from favorites!"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, user not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Authentication required"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found, favorite article not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Favorite article not found"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Internal Server Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}